refactor(auth): déplace la couche pure d'appartenance vers utils/membership
